+++ src/modules/customers/controllers/CustomerController.js
@@ -10,6 +10,7 @@ module.exports = () => ({
   create: async (req, res) => {
     try {
       const customer = await CreateCustomerBusiness().execute(req.body);
+      
       res.status(201).json(customer);
     } catch (error) {
       res.status(error.status || 500).json({
@@ -21,6 +22,7 @@ module.exports = () => ({
   findAll: async (_, res) => {
     try {
       const customers = await FindAllCustomerBusiness().execute();
+
       res.status(200).json(customers);
     } catch (error) {
       res.status(error.status || 500).json({
@@ -46,10 +48,12 @@ module.exports = () => ({
   update: async (req, res) => {
     try {
       const { id } = req.params;
+
       const editedCustomer = await UpdateCustomerBusiness().execute(
         id,
         req.body
       );
+
       res.status(200).json(editedCustomer);
     } catch (error) {
       res.status(error.status || 500).json({
@@ -61,7 +65,9 @@ module.exports = () => ({
   delete: async (req, res) => {
     try {
       const { id } = req.params;
+
       const deletedCustomer = await DeleteCustomerBusiness().execute(id);
+
       res.status(200).json(deletedCustomer);
     } catch (error) {
       res.status(error.status || 500).json({