+++ src/modules/routes/routes.js
@@ -20,21 +20,22 @@ const authController = AuthController();
 // Customer
 routes.post("/customers", validateToken, customerController.create);
 routes.get("/customers", customerController.findAll);
-routes.get("/customers/:id", customerController.findOne);
+routes.get("/customers/:id", validateId, customerController.findOne);
 routes.put("/customers/:id", validateToken, validateId, customerController.update);
 routes.delete("/customers/:id", validateToken, validateId, customerController.delete);
 
 // Order
 routes.post("/orders", validateToken, orderController.create);
 routes.get("/orders", orderController.findAll);
-routes.get("/orders/:id", orderController.findOne);
-routes.put("/orders/:id", validateToken, orderController.update);
+routes.get("/orders/:id", validateId, orderController.findOne);
+routes.put("/orders/:id", validateToken, validateId, orderController.update);
 routes.delete("/orders/:id", validateToken, validateId, orderController.delete);
 
 // User
-routes.post("/users/register", userController.create);
+routes.post("/users/register", validateToken, userController.create);
 routes.get("/users", userController.findAll);
 routes.get("/users/:id", validateId, userController.findOne);
+routes.delete("/users/:id", validateId, userController.delete);
 
 // Auth
 routes.post("/login", authController.create);