+++ src/modules/users/business/DeleteUserBusiness.js
@@ -1,0 +1,11 @@
+const { FindOneUserService, DeleteUserService } = require("../services");
+
+module.exports = () => ({
+  execute: async (id) => {
+    const user = await FindOneUserService().execute(id);
+
+    if (!user) throw { status: 404, message: "User not found" };
+
+    return await DeleteUserService().execute(id);
+  },
+});