+++ src/modules/users/services/index.js
@@ -1,4 +1,5 @@
 const CreateUserService = require("./CreateUserService");
+const DeleteUserService = require("./DeleteUserService");
 const FindAllUserService = require("./FindAllUserService");
 const FindOneUserService = require("./FindOneUserService");
 
@@ -6,4 +7,5 @@ module.exports = {
   CreateUserService,
   FindAllUserService,
   FindOneUserService,
+  DeleteUserService,
 };