+++ CHANGELOG.md
@@ -5,6 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.2] - 2023-07-10 - EfRxJxEW
+
+### Added
+
+- Update swagger docs
+
+## [2.3.1] - 2023-07-10 - EfRxJxEW
+
+### Added
+
+- Update swagger docs
+
+## [2.3.0] - 2023-07-09 - EfRxJxEW
+
+### Added
+
+- Update swagger docs
+
+## [2.2.1] - 2023-07-08 - 1qABvI3s
+
+### Added
+
+- Create swagger authentication
+
+## [2.2.0] - 2023-07-07 - LOOssm3S
+
+### Added
+
+- Create delete user functionality
 
 ## [2.1.0] - 2023-07-06 - 9xQcX25l